+++ src/extension.ts
@@ -6,6 +6,7 @@ import {
   window,
   workspace,
   ExtensionContext,
+  TextDocumentChangeEvent,
 } from "vscode";
 import tomlListener from "./toml/listener";
 import TomlCommands from "./toml/commands";
@@ -13,14 +14,19 @@ import TomlCommands from "./toml/commands";
 export function activate(context: ExtensionContext) {
   // Add active text editor listener and run once on start.
   context.subscriptions.push(window.onDidChangeActiveTextEditor(tomlListener));
-  workspace.onDidSaveTextDocument(() => {
-    tomlListener(window.activeTextEditor);
-  });
+
+  context.subscriptions.push(
+    workspace.onDidChangeTextDocument((e: TextDocumentChangeEvent) => {
+      const { fileName } = e.document;
+      if (fileName.toLocaleLowerCase().endsWith("cargo.toml")) {
+        tomlListener(window.activeTextEditor);
+      }
+    }),
+  );
   tomlListener(window.activeTextEditor);
 
   // Add commands
   context.subscriptions.push(TomlCommands.replaceVersion);
-    
 }
 
 export function deactivate() {}