+++ src/toml/commands.ts
@@ -3,12 +3,15 @@
  */
 import { commands, TextEditor, TextEditorEdit, Range } from "vscode";
 
+export const status = { inProgress: false };
+
 export const replaceVersion = commands.registerTextEditorCommand(
   "crates.replaceVersion",
   (editor: TextEditor, edit: TextEditorEdit, info) => {
-    if (editor && info) {
+    if (editor && info && !status.inProgress) {
       const { fileName } = editor.document;
       if (fileName.toLocaleLowerCase().endsWith("cargo.toml")) {
+        status.inProgress = true;
         console.log("Replacing", info.item);
         edit.replace(
           new Range(