+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to the "crates" extension will be documented in this file.
 
+### 0.0.10 
+* Complex Cargo.toml error fixed. [#1](https://github.com/serayuzgur/crates/issues/1)
+* Support for target dependencies. (not all usage cases)
+* Re-decorates on edit. (It was on save before)
+
 ### 0.0.9 
 * Version list (hover) is clickable now. Easier to navigate between versions.